+++ libs/vmm/examples/v1.5/shutdown-vz.sh
@@ -20,6 +20,7 @@ RUNNER_PID_FILE="${MOTLIE_VZ_RUNNER_PID_FILE:-$ARTIFACTS_DIR/${GUEST_NAME}-runne
 EGRESS_HELPER_PID_FILE="${MOTLIE_VZ_EGRESS_HELPER_PID_FILE:-$ARTIFACTS_DIR/${GUEST_NAME}-egress-helper.pid}"
 RUN_VM_DIR="$ARTIFACTS_DIR/${RUN_VM_NAME}.vm"
 EGRESS_SOCKET_PATH="${MOTLIE_VZ_EGRESS_SOCKET_PATH:-/tmp/motlie-vmm-${GUEST_NAME}.egress.sock}"
+EMBEDDED_EGRESS="${MOTLIE_VZ_EMBEDDED_EGRESS:-0}"
 
 terminate_pid_file() {
   local pid_file="$1"
@@ -44,6 +45,8 @@ terminate_pid_file() {
 }
 
 terminate_pid_file "$RUNNER_PID_FILE"
-terminate_pid_file "$EGRESS_HELPER_PID_FILE"
-rm -f "$EGRESS_SOCKET_PATH"
+if [[ "$EMBEDDED_EGRESS" != "1" ]]; then
+  terminate_pid_file "$EGRESS_HELPER_PID_FILE"
+  rm -f "$EGRESS_SOCKET_PATH"
+fi
 rm -rf "$RUN_VM_DIR"
